requirements.txt
aiohttp==3.9.1
aiohttp-proxy==0.1.2
aiohttp-socks==0.8.0
apscheduler==3.10.0
beautifulsoup4==4.12.2
fake-useragent==1.3.0
loguru==0.7.2
pydantic==2.5.2
python-dotenv==1.0.0
PyYAML==6.0.1
sqlalchemy==2.0.23